a load of fix

+++ JiayiLauncher/Features/Mods/Mod.cs
@@ -1,8 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Net.Http;
 using JiayiLauncher.Features.Game;
-using JiayiLauncher.Features.Launch;
 using JiayiLauncher.Utils;
 
 namespace JiayiLauncher.Features.Mods;
@@ -11,49 +11,52 @@ public class Mod
 {
 	public string Name { get; set; }
 	public string Path { get; set; }
+	public string MetadataPath { get; }
 	public List<string> SupportedVersions { get; set; }
 	public bool FromInternet => Path.StartsWith("http");
 
-	public Mod(string name, string path, List<string>? supportedVersions = null)
+	public Mod(string name, string path, List<string>? supportedVersions = null, string? metadataPath = null)
 	{
 		Name = name;
 		Path = path;
 		SupportedVersions = supportedVersions ?? new List<string> { "any version" };
+		MetadataPath = metadataPath ?? string.Empty;
 		
 		// just in case
-		if (!SupportedVersions.Contains("Any version")) return;
-		SupportedVersions.Remove("Any version");
-		SupportedVersions.Add("any version");
+		if (SupportedVersions.Contains("Any version"))
+		{
+			SupportedVersions.Remove("Any version");
+			SupportedVersions.Add("any version");
+		}
+		
+		if (!string.IsNullOrEmpty(MetadataPath)) return;
+
+		if (FromInternet)
+		{
+			var safeName = string.Join("", Name.Split(System.IO.Path.GetInvalidFileNameChars()));
+			MetadataPath = System.IO.Path.Combine(ModCollection.Current!.BasePath, ".jiayi", safeName + ".jmod");
+		}
+		else
+		{
+			var filename = System.IO.Path.GetFileNameWithoutExtension(Path);
+			var directory = System.IO.Path.GetDirectoryName(Path);
+			var modRelativePath = directory!.Replace(ModCollection.Current!.BasePath, string.Empty);
+			MetadataPath = System.IO.Path.Combine(ModCollection.Current.BasePath, ".jiayi", modRelativePath, filename + ".jmod");
+		}
 	}
 
-	public void SaveMetadata(ModCollection collection)
+	public void SaveMetadata()
 	{
 		if (SupportedVersions.Contains("Any version"))
 		{
 			SupportedVersions.Remove("Any version");
 			SupportedVersions.Add("any version");
 		}
-
-		string metadataPath;
 		
-		if (FromInternet)
-		{
-			// make sure the name doesn't contain any illegal characters
-			var safeName = string.Join("", Name.Split(System.IO.Path.GetInvalidFileNameChars()));
-			metadataPath = System.IO.Path.Combine(collection.BasePath, ".jiayi", safeName + ".jmod");
-		}
-		else
-		{
-			var filename = System.IO.Path.GetFileNameWithoutExtension(Path);
-			var directory = System.IO.Path.GetDirectoryName(Path);
-			var modRelativePath = directory!.Replace(collection.BasePath, string.Empty);
-			metadataPath = System.IO.Path.Combine(collection.BasePath, ".jiayi", modRelativePath, filename + ".jmod");
-		}
+		Directory.CreateDirectory(System.IO.Path.GetDirectoryName(MetadataPath)!);
+		File.WriteAllText(MetadataPath, $"{Name}\nat {Path}\nWorks on {string.Join(", ", SupportedVersions)}");
 		
-		Directory.CreateDirectory(System.IO.Path.GetDirectoryName(metadataPath)!);
-		File.WriteAllText(metadataPath, $"{Name}\nat {Path}\nWorks on {string.Join(", ", SupportedVersions)}");
-		
-		Log.Write("Mod.SaveMetadata()", $"{Name}'s metadata saved to {metadataPath}");
+		Log.Write("Mod.SaveMetadata()", $"{Name}'s metadata saved to {MetadataPath}");
 	}
 
 	public static Mod? LoadFromMetadata(string path)
@@ -66,8 +69,10 @@ public class Mod
 			var name = lines[0];
 			var modPath = lines[1].Replace("at ", string.Empty);
 			var supportedVersions = lines[2].Replace("Works on ", string.Empty).Split(", ").ToList();
+
+			var mod = new Mod(name, modPath, supportedVersions, path);
 			
-			return new Mod(name, modPath, supportedVersions);
+			return mod;
 		}
 		catch
 		{
@@ -78,34 +83,40 @@ public class Mod
 	
 	public void Delete(ModCollection collection)
 	{
-		string metadataPath;
+		if (!FromInternet && File.Exists(Path)) File.Delete(Path);
 		
-		if (FromInternet)
-		{
-			// make sure the name doesn't contain any illegal characters
-			var safeName = string.Join("", Name.Split(System.IO.Path.GetInvalidFileNameChars()));
-			metadataPath = System.IO.Path.Combine(collection.BasePath, ".jiayi", safeName + ".jmod");
-		}
-		else
-		{
-			var filename = System.IO.Path.GetFileNameWithoutExtension(Path);
-			var directory = System.IO.Path.GetDirectoryName(Path);
-			var modRelativePath = directory!.Replace(collection.BasePath, string.Empty);
-			metadataPath = System.IO.Path.Combine(collection.BasePath, ".jiayi", modRelativePath, filename + ".jmod");
-			
-			// also
-			File.Delete(Path);
-		}
-		
-		File.Delete(metadataPath);
+		File.Delete(MetadataPath);
 		collection.Mods.Remove(this);
 		Minecraft.ModsLoaded.Remove(this);
 		
 		Log.Write("Mod.Delete()", $"Deleted {Name} from the mod collection.");
 	}
+	
+	// for speed
+	private HttpResponseMessage? _response;
 
 	public bool IsValid()
 	{
+		if (_response != null) return _response.IsSuccessStatusCode;
+
+		if (FromInternet)
+		{
+			// ping
+			using var client = new HttpClient();
+			using var request = new HttpRequestMessage(HttpMethod.Head, Path);
+			
+			try
+			{
+				using var response = client.Send(request);
+				_response = response;
+				return response.IsSuccessStatusCode;
+			}
+			catch
+			{
+				return false;
+			}
+		}
+		
 		var exists = File.Exists(Path);
 		var isValidFile = Path.EndsWith(".dll") || Path.EndsWith(".exe");
         return exists && isValidFile;
